config: don't implicitly load coreos-ci-lib
Having it loaded implicitly is neat, though it has a major issue:
currently, Jenkins provides no way to avoid loading it, or overriding it
with a different git repo. This is a very important use case, because we
need to be able to test modifications to the library before merging. See
for example here where we want to hook up CI to coreos-ci-lib itself:

https://github.com/coreos/coreos-ci-lib/pull/15

(And correspondingly, on other repos which want to test their CI
combined with pending coreos-ci-lib changes).

Making it not implicitly loaded means that pipelines will now have to do
`@Library('coreos')` which isn't the end of the world.