jobs/seed-github-ci: stop wiping workspace
Otherwise we blow away the git repo which Jenkins uses to know whether
something changed, and so we end up rebuilding all the PRs and branches
every time a branch index refresh happens (currently scheduled for once
a day).

This essentially reverts #18. The original PR which prompted this is
merged now. Let's keep an eye on the upstream issue.