jobs/seed-github-ci: don't periodically scan branches
Because we don't have persistent storage, everytime Jenkins gets
reprovisioned, it wants to test ancient PRs when it does its first
scheduled repo scan. This unnecessarily bogs down the cluster and
increases the flakiness of test runs we actually care about.

So tell Jenkins to stop doing that.

The downside here is that Jenkins doesn't get to GC deleted
PRs/branches, but meh... the whole pod gets periodically wiped anyway.