jenkins: increase HEARTBEAT_CHECK_INTERVAL to 10 mins
The durable-task-plugin sometimes errors with "process apparently never
started" if the cluster is under heavy load because the script hasn't
had a chance to start yet. The default timeout is 5 mins, but let's try
to bump that to 10 mins to see if it helps. It may be that the problem
is elsewhere, in which case we can revert this.

Closes: #28 (hopefully)