Move to local PVC
After discussions with the CentOS CI folks, I was pointed to an
alternative to the NFS PVC. We can use a "local PVC" which is tied to a
specific node.

This does mean that Jenkins itself will be pinned to that node (but
anyway, all the heavy work is done in separate unconstrained pods) and
that we'll lose all the data if a hardware failure happens (there's no
redundancy/backups like NFS).

This is fine for upstream CI though and is much better than what we
currently have.