jenkins/config: workaround jcasc bug in openshift/jenkins
There is a bug in the Jenkins OpenShift image which causes the
Kubernetes credential that it was supposed to create not to happen if
there are JCASC dropins which also create credentials, like we do here.
The symptom is:

    org.jenkinsci.plugins.kubernetes.auth.KubernetesAuthException: No
    credentials found with id 1a12dfa4-7fc5-47a7-aa17-cc56572a41c7

Use the documented workaround to fix this.

See: https://bugzilla.redhat.com/show_bug.cgi?id=1956652
See: https://access.redhat.com/solutions/6016001