Stop mounting GitHub OAuth secret into Jenkins
Just have it live directly in the configmap. This is not as elegant as
the kubernetes-credentials-provider but takes us a step closer to
dropping `jenkins.yaml`.

There isn't really a difference today in terms of RBAC in our namespace
between secrets and configmaps (e.g. all authenticated users with access
to the namespace can access both, and so can the `jenkins` SA).