jobs: add `test-override` job
This job is similar in structure to `bump-lockfile` (i.e. it runs tests
on multiple arches in parallel), but is geared towards testing Bodhi
and/or Koji overrides. The `OVERRIDES` parameter supports a list of URLs
from those services that the job will automatically download into the
`overrides/rpm` directory when doing the build.

For now, like bump-lockfile, this only does local testing, i.e. QEMU and
testiso scenarios. We could extend this in the future to do some cloud
testing too.

This is prep for having another job that triggers this one based on
Bodhi Fedora messages, but it's structured to be useful on its own too
for ad-hoc testing.