jobs: add `bodhi-trigger` job
This job is a step towards better integrating our CI into the rest of
Fedora QA.

The job triggers on Fedora messages from Bodhi whenever an update is
created or edited. It then resolves the Bodhi update to a relevant FCOS
stream, and trigger the `test-override` job with the right parameters.

The implementation details here are a bit messy. Triggering on "a new
Bodhi update" isn't as simple as I would've liked and mapping back to
streams is a bit hairy, but it works!

There's of course a lot of cleanups we could do here (many are marked
in the code), though I'm hoping we can get this in to gain some early
experience and work on cleanups in parallel with the larger task of
adding reporting and socializing it.

See also: https://github.com/coreos/fedora-coreos-tracker/issues/17