jobs/test-override: on mechanical streams, build on top of known good version
When testing a rawhide update, we should be testing on top of a rawhide
build that we know has worked. Otherwise, the test may fail for any
arbitrary reason. Since rawhide is not locked, we do this by autolocking
on top of the most recent build that contains all the relevant arches.

See also: https://github.com/coreos/fedora-coreos-releng-automation/issues/181