jobs/bodhi-trigger: read SRPM list from `bodhi-testing.yaml`
Add a new `bodhi-testing.yaml` file containing the list of SRPMs to
test. We'll augment the schema of this file later on to allow e.g.
running only a subset of tests.

This will also in the short-term serve as the source of truth for which
packages are gated on these tests or not. Long-term, we may use this
file slightly differently, though likely the tests to run per component
will still live there (including the default tests to run on any
unlisted component).

While we're here, add a few more core packages to the list.