bodhi-testing.yaml: support specifying tests to run
For some packages, it's silly to run the full testsuite when they're not
involved or tested at all in e.g. the testiso scenarios.

Let's allow configuring the list of tests to run per SRPM.

This is also prep for being able to run e.g. just `basic` for a much
wider set of packages.