bodhi-testing.yaml: make test patterns more globby
The test pattern for `toolbox` was incorrect and so didn't match
anything.

For that component and some others, instead of specifying a full test
path, make the patterns more globby. That makes them more obviously
right, and resistant to the tests moving.

Fixes 6c1f869 ("bodhi-testing.yaml: support specifying tests to run").