jobs/bodhi-trigger: report SUCCESS even for UNSTABLE builds
When a snoozed test expires, if it has the `warn: true` key on, we'll
start warning and that'll show up as an UNSTABLE build. These warnings
are more for test maintainers than for public consumption so don't
translate them to NEEDS_INSPECTION because that will block updates on
gated components.