jobs/bodhi-trigger: respect `test-patterns: skip`
If a test pattern is marked as `skip`, then respect that. Don't force
`basic` to run in that case. The logic there was meant as a way to force
`basic` to always run *if* some other tests were also selected, not to
counteract `skip`.

Fixes 6c1f869 ("bodhi-testing.yaml: support specifying tests to run").