jobs/bodhi-trigger: also report test successes
Sometimes, Bodhi testing is broken for a few days if rawhide is in a
bad shape. Let's also have the test report successes so that it's easier
to tell when upstream testing have been fixed. It makes it more obvious
also when e.g. just a single run failed in a series of otherwise passing
runs.