test-override: always buildfetch
Right now we only buildfetch if we are doing autolock (i.e. basing
on mechanical streams). Let's always buildfetch so we can see an
RPM diff when the commit/container is built.

Right now there is a bug where `cosa diff` when it's the first build
in the stream will fail (i.e. causing `cosa build` to fail), but should
be fixed soon. This works around that, but also shows the RPM diff,
which is nice too.